requirements.txt
selenium==3.141.0
pytest==6.2.5
allure-pytest==2.9.42
pytest-rerunfailures==10.0
faker==8.10.3
webdriver-manager==3.4.2

